third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGLinearGradientElement-svgdom-x1-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x1' property of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.x1.baseVal.valueAsString is "100%"
PASS linearGradientElement.x1.baseVal.valueAsString is "200%"
PASS successfullyParsed is true

TEST COMPLETE
